requirements.txt
openai==0.27.4
python-dotenv==0.19.2